Update README.md
Okay, this looks scary! :-)

I think this is a really important change. People visiting the REPO today does not really get an introduction to what Baobab is and especially why you would want to use it. The README has become too big.

So I created a WIKI that has a more "how to use" perspective, than a "how it is built" perspective. I think that is important to give developers a good intro. I also expand on this to give introductions to patterns and how to solve specific scenarios.

I also included the article I wrote as it has gotten really good feedback, even [dailyjs](http://dailyjs.com/) mentioned it being a good introduction to Baobab.

To me this LIB has been as big as the React JS release itself. It truly is the "next generation" of application architecture and we have to make it easy for other developers to understand what this is all about :-)

I have also reworked a lot of the examples to be more app related. I hope you like :-)