Conform behavior for splice's deleteCount to es6 spec
ES6 section 7.1.3 (ToNumber) specifies what should happen when the
argument for the nb (deleteCount) is not a number. The code now
correctly handles booleans, strings, decimals, null, and undefined when
passed in as that argument. As a side effect it gets rid of the use of
`Number.isNaN` and `Number.parseInt` which aren't suported in IE or
Safari.